+++ mod-data-import-converter-storage-server/src/main/java/org/folio/dao/MatchProfileDaoImpl.java
@@ -1,7 +1,6 @@
 package org.folio.dao;
 
 import io.vertx.core.Future;
-import io.vertx.ext.sql.SQLConnection;
 import org.folio.rest.jaxrs.model.MatchProfile;
 import org.folio.rest.jaxrs.model.MatchProfileCollection;
 import org.folio.rest.persist.interfaces.Results;
@@ -14,8 +13,6 @@ import org.springframework.stereotype.Component;
 public class MatchProfileDaoImpl extends AbstractProfileDao<MatchProfile, MatchProfileCollection> {
 
   private static final String MATCH_PROFILES_TABLE_NAME = "match_profiles";
-  private static final String MATCH_TO_ACTION_ASSOCIATION_TABLE = "match_to_action_profiles";
-  private static final String MATCH_TO_MATCH_ASSOCIATION_TABLE = "match_to_match_profiles";
 
   @Override
   String getTableName() {
@@ -42,11 +39,5 @@ public class MatchProfileDaoImpl extends AbstractProfileDao<MatchProfile, MatchP
   @Override
   protected MatchProfile markProfileEntityAsDeleted(MatchProfile profile) {
     return profile.withDeleted(true);
-  }
-
-  @Override
-  protected Future<Boolean> deleteAllAssociationsWithDetails(Future<SQLConnection> txConnection, String masterProfileId, String tenantId) {
-    return deleteAssociationsWithDetails(txConnection, masterProfileId, MATCH_TO_ACTION_ASSOCIATION_TABLE, tenantId)
-      .compose(voidDeleteRes -> deleteAssociationsWithDetails(txConnection, masterProfileId, MATCH_TO_MATCH_ASSOCIATION_TABLE, tenantId));
   }
 }