+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 2019-06-19 v1.4.0-SNAPSHOT
 * Added removing all profile associations when it's marked as deleted.
+* Changed implementation of delete functionality for profile associations using DB rule.
 
 ## 2019-06-12 v1.3.0
 * Updated association profiles samples.